Add remote.datatypes package. Move all datatype classes in remote to datatypes package. Add PIDParams datatype class for representing a collection of PIDParams.
