+++ src/remote/Server.java
@@ -4,6 +4,9 @@ import enums.ControlMode;
 import enums.PacketCommand;
 import exceptions.MissingIDException;
 import helpers.DataConversionHelper;
+import remote.datatypes.CommunicationPacket;
+import remote.datatypes.Error;
+import remote.datatypes.PacketList;
 import remote.listeners.AcknowledgementListener;
 import remote.listeners.ErrorListener;
 import remote.listeners.ExceptionListener;
