+++ src/remote/datatypes/PacketList.java
@@ -1,4 +1,4 @@
-package remote;
+package remote.datatypes;
 
 import enums.PacketCommand;
 import exceptions.IncorrectDataException;
